restructure status and watch actions
